Made CreateSessionTicket static Renamed CarID to CarId Changed Permission from int to Enum

+++ src/AuthServer/Util/AuthConsoleCommands.cs
@@ -44,7 +44,7 @@ namespace AuthServer.Util
                 return CommandResult.InvalidArgument;
             }
             
-            user.Permission = (int)perm;
+            user.Permission = perm;
             AccountModel.Update(AuthServer.Instance.Database.Connection, user);
             Log.Info("User {0} has now permission {1}!", user.Username, Enum.GetName(typeof(UserPermission), user.Permission));
             return CommandResult.Okay;
